backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
bcrypt
pyjwt
pytest
pytest-asyncio